requirements.txt
attrs==20.3.0
aws-cdk.assets==1.110.1
aws-cdk.aws-applicationautoscaling==1.110.1
aws-cdk.aws-autoscaling==1.110.1
aws-cdk.aws-autoscaling-common==1.110.1
aws-cdk.aws-certificatemanager==1.110.1
aws-cdk.aws-cloudformation==1.110.1
aws-cdk.aws-cloudfront==1.105.0
aws-cdk.aws-cloudwatch==1.110.1
aws-cdk.aws-codedeploy==1.110.1
aws-cdk.aws-codeguruprofiler==1.110.1
aws-cdk.aws-codestarnotifications==1.110.1
aws-cdk.aws-ec2==1.110.1
aws-cdk.aws-ecr==1.110.1
aws-cdk.aws-ecr-assets==1.110.1
aws-cdk.aws-efs==1.110.1
aws-cdk.aws-elasticloadbalancing==1.110.1
aws-cdk.aws-elasticloadbalancingv2==1.110.1
aws-cdk.aws-events==1.110.1
aws-cdk.aws-iam==1.110.1
aws-cdk.aws-kms==1.110.1
aws-cdk.aws-lambda==1.110.1
aws-cdk.aws-logs==1.110.1
aws-cdk.aws-rds==1.110.1
aws-cdk.aws-route53==1.110.1
aws-cdk.aws-s3==1.110.1
aws-cdk.aws-s3-assets==1.110.1
aws-cdk.aws-s3-deployment==1.105.0
aws-cdk.aws-sam==1.110.1
aws-cdk.aws-secretsmanager==1.110.1
aws-cdk.aws-signer==1.110.1
aws-cdk.aws-sns==1.110.1
aws-cdk.aws-sqs==1.110.1
aws-cdk.aws-ssm==1.110.1
aws-cdk.cloud-assembly-schema==1.110.1
aws-cdk.core==1.110.1
aws-cdk.custom-resources==1.110.1
aws-cdk.cx-api==1.110.1
aws-cdk.lambda-layer-awscli==1.105.0
aws-cdk.region-info==1.110.1
boto3==1.17.94
botocore==1.20.94
cattrs==1.5.0
-e git+https://git-codecommit.us-east-1.amazonaws.com/v1/repos/CDK-3TA@46c6fab537e1d746a936538b3fc24a68f82cb203#egg=cdk_3_ta
click==8.0.1
constructs==3.3.93
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsii==1.30.0
MarkupSafe==2.0.1
publication==0.0.3
PyMySQL==1.0.2
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
